chore: upgrade kafkaflow.sample.flowcontrol to modern c# practices
