+++ samples/KafkaFlow.Sample.FlowControl/SampleMessage.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace KafkaFlow.Sample.FlowControl
+{
+    internal class SampleMessage
+    {
+        public Guid Id { get; set; }
+
+        public string Name { get; set; }
+    }
+}
